+++ pennylane_qiskit/converter.py
@@ -129,7 +129,7 @@ def execute_supported_operation(operation_name: str, parameters: list, wires: li
     """Utility function that executes an operation that is natively supported by PennyLane.
 
     Args:
-        operation_name (str): wires specified for the template
+        operation_name (str): Name of the PL operator to be executed
         parameters (str): parameters of the operation that will be executed
         wires (list): wires of the operation
     """
@@ -137,7 +137,7 @@ def execute_supported_operation(operation_name: str, parameters: list, wires: li
 
     if not parameters:
         operation(wires=wires)
-    elif operation_name == "QubitStateVector":
+    elif operation_name in ["QubitStateVector", "StatePrep"]:
         operation(np.array(parameters), wires=wires)
     else:
         operation(*parameters, wires=wires)
@@ -179,7 +179,8 @@ def load(quantum_circuit: QuantumCircuit):
 
         # Processing the dictionary of parameters passed
         for op, qargs, cargs in qc.data:
-            instruction_name = op.__class__.__name__
+            # the new Singleton classes have different names than the objects they represent, but base_class.__name__ still matches
+            instruction_name = getattr(op, "base_class", op.__class__).__name__
 
             operation_wires = [wire_map[hash(qubit)] for qubit in qargs]